requirements.txt
yfinance==0.2.28
schedule==1.2.1
python-dotenv==1.0.0
requests==2.31.0
pytz==2023.3
